requirements.txt
pandas
numpy
scikit-learn
torch
torchvision
streamlit
requests
beautifulsoup4
selenium
sqlalchemy
psycopg2-binary
ipykernel
opencv-python
python-dotenv
PyMySQL
webdriver-manager
curl_cffi
tqdm